refactor: use load function in auth pages
Instead of using hacky opacity trick, simply load before showing
