+++ src/routes/(auth)/signin/+page.ts
@@ -1,0 +1,19 @@
+import { goto } from "$app/navigation";
+import { backendService } from "$lib/grpc-api";
+import { AuthenticationStatus } from "$lib/model/api/authentication";
+import { Nothing } from "$lib/model/api/base";
+import type { PageLoad } from "./$types";
+
+export const load: PageLoad = async () => {
+    try {
+        const authStatus = (await backendService.getAuthenticationStatus(Nothing).response)
+            .authenticationStatus;
+        if (authStatus === AuthenticationStatus.AUTHENTICATED) {
+            await goto("/");
+        }
+    } catch (error) {
+        console.error("There was an error acquiring the authentication status:", error);
+    }
+
+    return {};
+};
